==> scr/00_libraries.R <==
renv::install(c("tidyverse","Seurat","finalfit"))

renv::install("bioc::miloR")

renv::install("phipsonlab/speckle")
renv::install("statmod")
renv::install("ggalluvial")

renv::install(c("coda.base", "psych"))
renv::install("kharchenkolab/sccore")
renv::install("kharchenkolab/conos")
renv::install("kharchenkolab/cacoa")
renv::install("rpolicastro/scProportionTest")


renv::install("bioc::scater")
renv::install("bioc::scran")

renv::install("bioc::ensembldb")
renv::install("bioc::AnnotationDbi") 
renv::install("bioc::AnnotationHub")

renv::install("ggbeeswarm")

renv::install("harmony")

renv::install("circlize")

renv::install("bioc::MouseThymusAgeing")
renv::install("satijalab/seurat-data")
